Fix Install Name of Joomla
Summary:
We got the update from the Joomla store. It says that:"The install name of your extension "Facebook Pixel" doesn't match the listing name "Official Facebook Pixel". Please update the name in your XML/language files or request for the listing name to be changed."

So make a change according to their feedback.

Reviewed By: g233007

Differential Revision: D13308331

fbshipit-source-id: 4592a5f9